+++ cmds/test.go
@@ -56,9 +56,9 @@ func doTest(cmd *cobra.Command, args []string) {
 		}
 
 		result := resultOutput(test, "")
-		successCount = +result.SuccessCount
-		failCount = +result.FailCount
-		totalCount = +result.TotalCount
+		successCount = result.SuccessCount
+		failCount = result.FailCount
+		totalCount = result.TotalCount
 
 		if result.ExitCode == 1 {
 			exitCode = 1