+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 0.12.4
 - Re tweak agent header
+- Fix bug with emtpy response body 
+
 0.12.3
 - Fix issues gofakeit date related functions
 - Fix agent header formatting